pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "framed-cli"
version = "0.1.0"
description = "Automated App Store screenshot tool for iOS"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "yamshta" },
]
dependencies = [
  "Pillow>=10.0.0",
  "PyYAML>=6.0",
  "click>=8.0.0",
]

[project.scripts]
framed = "framed.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/framed"]
